+++ topobank/users/management/commands/notify_users.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 from django.contrib.sites.models import Site  # we use this as object which is always there, this is a workaround
-from django.contrib.staticfiles.templatetags.staticfiles import static
+from django.conf.urls.static import static
 from django.shortcuts import reverse
 
 from notifications.signals import notify
@@ -14,6 +14,7 @@ _log = logging.getLogger(__name__)
 
 
 class Command(BaseCommand):
+    """Command for sending notifications to users."""
     help = """Send a notification to all or a bunch of users.
 
     The recipients can read the notification when logged in.